Рефакторит форму загрузки изображения
